\name{cmos}
\alias{cmos}
\docType{data}
\title{Sample Dataset}
\description{A sample dataset with 3 patients, 4 variables,14 records.}
\usage{data(cmos)}
\format{
  A data frame with 14 observations on the following 4 variables.
  \describe{
    \item{\code{ptid}}{a numeric vector}
    \item{\code{rxdate}}{a character vector, prescription refill date}
    \item{\code{supplies}}{a numeric vector}
    \item{\code{rxDay}}{a numeric vector, prescription refill date}
  }
}

\examples{ 
data(cmos)
cmos
}
\keyword{datasets}
